Remove unused edgefind variables, shuffle utils around
