+++ util/processing.py
@@ -1,29 +1,10 @@
 import numpy as np
-from dataclasses import dataclass
 import cv2
 from util.logger import logger
 
 
 RGB = list[int]
 FlakeRGB = np.ndarray[int]
-
-
-@dataclass
-class Box:
-    label: str
-    x: int
-    y: int
-    width: int
-    height: int
-
-    def to_mask(self, img, b=5):
-        h, w = img.shape
-        bound_x = min(0, self.x)
-        bound_y = min(0, self.y)
-        return np.logical_and.outer(
-            np.logical_and(np.arange(bound_y, h) >= self.y - b, np.arange(bound_y, h) <= self.y + self.height + 2 * b),
-            np.logical_and(np.arange(bound_x, w) >= self.x - b, np.arange(bound_x, w) <= self.x + self.width + 2 * b),
-        )
 
 
 def bg_to_flake_color(rgb: RGB) -> FlakeRGB:
@@ -45,21 +26,17 @@ def bg_to_flake_color(rgb: RGB) -> FlakeRGB:
 def edgefind(imchunk: np.ndarray, avg_rgb: FlakeRGB, pixcals: list[float], t_rgb_dist: int) -> tuple[RGB, any, float]:  # TODO
     """
     TODO
-    :param imchunk: The pixels to find edges in.
+    :param imchunk: The image chunk to find edges in.
     :param avg_rgb: The average flake RGB, from `bg_to_flake_color()`.
     :param pixcals:
     :param t_rgb_dist: The threshold a pixel color must be within from the average flake color to be counted as good.
     :return: The results, as a tuple of (flake rgb, edge image, flake area).
     """
     pixcalw, pixcalh = pixcals
-    edgerad = 20
-
-    imchunk2 = imchunk.copy()
-    impix = imchunk.copy().reshape(-1, 3)
     dims = np.shape(imchunk)
 
+    impix = imchunk.copy().reshape(-1, 3)
     flakeid = np.sqrt(np.sum((impix - avg_rgb) ** 2, axis=1)) < t_rgb_dist  # a mask for pixel color
-    maskpic = np.reshape(flakeid, (dims[0], dims[1], 1))
 
     red_freq = np.bincount(impix[:, 0] * flakeid)
     green_freq = np.bincount(impix[:, 1] * flakeid)
@@ -83,44 +60,3 @@ def edgefind(imchunk: np.ndarray, avg_rgb: FlakeRGB, pixcals: list[float], t_rgb
                .astype(np.int16) * np.array([25, 25, 25]) / 255
 
     return rgb, edgeim.astype(np.uint8), farea
-
-
-def merge_boxes(dbscan_img, boxes: list[Box], eliminated_indexes: list[int] = []) -> list[Box]:
-    """
-    Merges a list of boxes by combining boxes with overlap.
-    :param dbscan_img: TODO
-    :param boxes: The list of boxes to merge.
-    :param eliminated_indexes: TODO
-    :return: The merged list.
-    """
-    merged = []
-
-    for _i in range(len(boxes)):
-        if _i in eliminated_indexes:
-            continue
-        i = boxes[_i]
-        for _j in range(_i + 1, len(boxes)):
-            j = boxes[_j]
-            # Ith box is always <= jth box regarding y. Not necessarily w.r.t x.
-            # sequence the y layers.
-            # just cheat and use Intersection in pixel space method.
-            on_i = i.to_mask(dbscan_img)
-            on_j = j.to_mask(dbscan_img)
-
-            # Now calculate their intersection. If there's any overlap we'll count that.
-            intersection_count = np.logical_and(on_i, on_j).sum()
-
-            if intersection_count > 0:
-                # Extend the first box to include dimensions of the 2nd box.
-                x_min = min(i.x, j.x)
-                x_max = max(i.x + i.width, j.x + j.width)
-                y_min = min(i.y, j.y)
-                y_max = max(i.y + i.height, j.y + j.height)
-
-                new_width = x_max - x_min
-                new_height = y_max - y_min
-                i = Box(i.label, x_min, y_min, new_width, new_height)
-                eliminated_indexes.append(_j)
-        merged.append(i)
-
-    return merged
